requirements.txt
Flask
SQLAlchemy
python-dotenv
alembic
flask-cors
psycopg[binary]
gunicorn
librosa
numpy
scipy
pydub
torch
audioread
requests
matplotlib